% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{hallmark_cat}
\alias{hallmark_cat}
\title{Functional categories of HALLMARK gene sets
Derived from Liberzon, Cell Systems, 2015, PMID: 26771021}
\description{
Functional categories of HALLMARK gene sets
Derived from Liberzon, Cell Systems, 2015, PMID: 26771021
}
\author{
HCM
}
\keyword{data}
