% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{hpa}
\alias{hpa}
\title{Gene annotation from the Human Protein Atlas}
\format{
tibble
}
\description{
Information on genes, e.g. description, protein class or chromosome. Useful
for annotating Differential Gene Expression results
}
\author{
HCM
}
\keyword{data}
