% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{msigdb}
\alias{msigdb}
\title{MSigDb v7.2
Downloaded from the Broad Institute on November 2020}
\format{
List with 25 elements containing human gene symbols
\describe{
\item{c1.all}{chromosomal position gene sets, n = 299}
\item{c2.all}{canonical pathways, n = 6229}
\item{c2.cgp}{chemical and genetic pertubations, n = 3358}
\item{c2.cp.biocarta}{biocarta pathways, n = 292}
\item{c2.cp.kegg}{kegg pathways, n = 186}
\item{c2.cp.pid}{pid pathways, n = 196}
\item{c2.cp.reactome}{reactome pathways, n = 1554}
\item{c2.cp}{canonical pathways, n = 2871}
\item{c2.cp.wikipathways}{wiki pathways, n = 587}
\item{c3.all}{micro RNA and transcription factor targets, n = 3556}
\item{c3.mir.mirdb}{micro RNA targets, n = 2377}
\item{c3.mir}{micro RNA targets, n = 2598}
\item{c3.tft.gtrd}{trancription factor targets, n = 348}
\item{c3.tft}{trancription factor targets, n = 958}
\item{c4.all}{cancer gene neighborhoods and cancer modules, n = 858}
\item{c4.cgn}{cancer gene neighborhoods, n = 427}
\item{c4.cm}{cancer modules factor, n = 431}
\item{c5.bp}{GO biological processes, n = 4436}
\item{c5.cc}{GO cellular components, n = 580}
\item{c5.mf}{GO molecular function, n = 901}
\item{c5.hpo}{Human Phenotype Ontology, n = 4494}
\item{c6.all}{oncogenic signatures, n = 189}
\item{c7.all}{immunologic signatures, n = 4872}
\item{c8.all}{cell type signatures, n = 302}
\item{h.all}{hallmark gene sets, n = 50}
}
}
\description{
MSigDb v7.2
Downloaded from the Broad Institute on November 2020
}
\author{
HCM
}
\keyword{data}
