% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{pda_regulons}
\alias{pda_regulons}
\title{ARACNe regulons derived from a) 242 CUMC pancreatic specimen (PanIN/IPMN/PDA) and b) 426 ICGC
Canada PDA specimen (PDA primaries and metastasis)}
\format{
list of regulons, each comprising one regulon with its RP
}
\description{
ARACNe regulons derived from a) 242 CUMC pancreatic specimen (PanIN/IPMN/PDA) and b) 426 ICGC
Canada PDA specimen (PDA primaries and metastasis)
}
\author{
HCM
}
\keyword{data}
